Create new view utils package, move ViewHolder
